Extract `timeAgo` to its own file
