+++ source/features/repo-age.tsx
@@ -1,6 +1,7 @@
 import React from 'dom-chef';
 import select from 'select-dom';
 import * as api from '../libs/api';
+import timeAgo from '../libs/time-ago';
 import features from '../libs/features';
 import * as icons from '../libs/icons';
 import {getRepoGQL} from '../libs/utils';
@@ -13,18 +14,6 @@ async function getRepoCreationDate(): Promise<Date> {
 	`);
 
 	return new Date(repository.createdAt);
-}
-
-function timeAgo(date: Date): {interval: number; timespan: string} {
-	const units = ['minute', 'hour', 'day', 'year'];
-
-	const ago = (<time-ago datetime={date.toISOString()} format="micro"/>).textContent;
-	const [interval, short] = ago!.match(/[a-z]+|[^a-z]+/gi)! as [number, string];
-
-	let timespan = units.find(unit => unit.startsWith(short))!;
-	timespan = interval > 1 ? timespan + 's' : timespan;
-
-	return {interval, timespan};
 }
 
 async function init(): Promise<void> {
